requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4

# Google Cloud
google-cloud-secret-manager==2.22.0
google-cloud-storage==2.19.0
google-cloud-firestore==2.19.0
google-cloud-pubsub==2.27.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.157.0
